+++ packages/frontier-evm/src/index.ts
@@ -320,17 +320,26 @@ const CallProcessor: SecondLayerHandlerProcessor_1_0_0<
       ? (tx as TransactionV2).asLegacy
       : (tx as EthTransaction);
 
-    let from, hash, to, success;
+    let from = '',
+      hash = '',
+      to,
+      success;
     try {
       const executionEvent = getExecutionEvent(original);
       from = executionEvent.from;
-      to = executionEvent.to;
+      // There is no test for tx being of type EthTransaction
+      if (!(rawTx as EIP1559Transaction).action || !(rawTx as EIP1559Transaction).action.isCreate) {
+        to = executionEvent.to;
+      }
       hash = executionEvent.hash;
       success = executionEvent.status.isSucceed;
     } catch (e) {
+      logger.warn(
+        `Unable to get executionEvent for call. block='${original.block.block.header.number.toNumber()}', index='${
+          original.idx
+        }'`
+      );
       success = false;
-
-      throw new Error('Unable to get executionEvent for transaction');
     }
 
     let call: FrontierEvmCall;