+++ packages/frontier-evm/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.4] - 2023-05-29
+### Fixed
+- `to` field being defined on contract creation. It is now undefined (#51)
+
+## [0.1.3] - 2023-05-10
+### Fixed
+- Don't throw when not able to get execution event (#49)
+
 ## [0.1.2] - 2022-08-26
 No changes, fix release skipped.